add(auth): try with silent authentication when sso is enabled
In case that `isauth` responds with not authenticated and with sso enabled, the first attempt will
be to authenticate silently. It will call the SSO url (saml) within the iFrame and check again. In
case that authentication succeeded and successfully set cookies, it should respond as authenticated
and continue with further actions.